Pass test_comment as an empty string on test run
The test_comment should be either a string or an empty string. `None` is not
being accepted as a valid value anymore.

Also print more information when an error is found when adding a test record
for a test case. This will make easy to identify what was the test case for the
failure since we are running things in a multiprocessing environment.