Added new `test-plan` command.
Added a new `test-plan` command which allows one to create a new Test Plan
into Polarion. There are two possible types of Test Runs one can create:
Releases or Iterations. Furthermore, one can create a new Test Plan as a child
of an existing Test Plan (regardless of their types).

This new command will first query for a Test Run that matches the passed `name`
argument. If a match is found, then an instance of a `Test Plan` will be
returned. Otherwise, a new Test Plan will be created and returned.