Drop Testimony usage as a library
Testimony was being used basically to walk a given path and collect the
test methods and functions it founds. Also it was used to collect the
fields from tests' docstrings. The output provided by Testimony used to
be plain test formatted as reStructuredText and therefore Betelgeuse had
to implement its own RST parser to convert the field's value into HTML
to be imported into Polarion.

Testimony continues to be a cli tool which produces text based output
while Betelgeuse is interested on having the HTML output. As Betelgeuse
evolved it needed to support more complex docstrings and using Testimony
would require to create a reverse parser from HTML to text (RST
formatted) to later on Betelgeuse parse that RST into HTML.

With all that said the decision to drop Testimony usage is just because
will be easier to parse the docstring and get the HTML output then to
create that reverse parser. Also the test case collection can be
simplified and improved for Betelgeuse specific use without requiring a
lot of work on Testimony project.

With this change, test case collection tools are located on the
`collector` module while parsing tools are located on `parser` module.