Rename xml-<command>s to <command>s
Creating/updating using Pylarion is deprecated in favor of the XML
importers so the xml-test-case and xml-test-run commands are now renamed
to test-case and test-run respectively and the old test-case and
test-run commands are dropped.

Provide aliases for xml-test-case and xml-test-run to test-case and
test-run respectively to not break Betelgeuse for users using those
commands. These aliases are going to be dropped in a future version of
Betelgeuse.