Allow providing a default value for all fields
Improve create_xml_testcase in order to make it look for a default
values for all fields, not  only the custom fields.

Also improve documentation to reflect the changes and expose the field
available for the test case object passed when a callable is used as the
default value generator.

Closes #106